Add two functions that help with date formatting
getDayName() function takes a date string and uses the date object to
calculate the week day number (from 0 to 6) and returns the day's name
from a predefined weekday names array.

getDayAndMonth() function takes a date string and uses the date object
to calculate the month number (from 0 to 11) and adds the month's name
from a predefined month manes array to the day of the month.

return format: [day] [month name]