Add an event for when user searches for location
the event runs when user writes something into search box. the event
then waits for 500ms and then rechecks that user didn't input anything
else. in that case the user stopped typing and so the event invokes the
function responsible for retrieving and displaying weather data.

NOTE: if search box is empty, search value is set to cairo