Create folder for contact page
why a separate folder for contact page?
since servers use index.html as the default page for a folder
creating a folder named contact with index.html file inside it
means that if i reference the folder, the browser will open the
index.html.

this is useful for removing the weird looking html extension at
the end of each url. this way only the directory name will show